Update keyboard forcus before computing pressed keys
Prior to this, the first keypress would trigger keyboard focus and the
focus would get computed after we figured out the key was down.  The result
was that the client got sent the key as part of the array of keys sent on
enter rather than as an actual wl_keyboard.key event.